requirements_scripts.txt
wandb
torchsummary
lightning
